+++ webcaf/webcaf/caf/util.py
@@ -1,9 +1,4 @@
-from abc import abstractmethod
 from typing import Any, Dict, Optional
-
-from django.contrib.auth.mixins import LoginRequiredMixin
-
-from webcaf.webcaf.views.session_utils import SessionUtil
 
 
 class IndicatorStatusChecker:
@@ -68,22 +63,3 @@ Please confirm you agree with this status, or you can choose to change the outco
             "override_status": override_status,
             "outcome_status_text": outcome_status_text_map[outcome_status],
         }
-
-
-class UserRoleCheckMixin(LoginRequiredMixin):
-    def dispatch(self, request, *args, **kwargs):
-        if not request.user.is_authenticated:
-            return self.handle_no_permission()
-        else:
-            user_profile = SessionUtil.get_current_user_profile(request)
-            if user_profile.role not in self.get_allowed_roles():
-                return self.handle_no_permission()
-            return super().dispatch(request, *args, **kwargs)
-
-    @abstractmethod
-    def get_allowed_roles(self) -> list[str]:
-        """
-        Needs to be implemented by the subclass.
-        List of roles that are allowed to access the view.
-        :return:
-        """
