+++ webcaf/webcaf/templates/assessment-profile.html
@@ -8,7 +8,7 @@
                 Government CAF profile
             </h1>
             <p>
-                Every system undergoing GovAssure must be classified as one of two profile types. These are baseline or
+                Every system undergoing an assessment must be classified as one of two profile types. These are baseline or
                 enhanced.
             </p>
             <div class="govuk-inset-text">
