+++ webcaf/webcaf/templates/system/system-details.html
