--- webcaf/webcaf/templates/title.html
@@ -1,37 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% load govuk_frontend_django %}
-{% load csp %}
-
-{% block title %}{{ title }}{% endblock %}
-
-{% block content %}
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-two-thirds">
-            <h2 class="govuk-heading-l">
-                {{ title }}
-            </h2>
-            <p class="govuk-body">
-                {{ description }}
-            </p>
-        </div>
-    </div>
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-two-thirds">
-            <form method="post">
-                {% csrf_token %}
-                <button name="submit" class="govuk-button govuk-button" id="id_submit" data-module="govuk-button"
-                        data-govuk-button-init="">Continue
-                </button>
-            </form>
-
-        </div>
-    </div>
-{% endblock %}
-
-{% block bottomjs %}
-
-{% endblock %}
-
-{% block extra_stylesheets %}
-{% endblock %}
